Move `DebugWithDepth` to `param_types.rs`
