+++ packages/fuels-core/src/utils.rs
@@ -1,5 +1,2 @@
-mod debug_with_depth;
-pub(crate) use debug_with_depth::DebugWithDepth;
-
 pub mod constants;
 pub mod offsets;
